Add QFileEncrypted
This allows easy encryption support in anything that uses
QFile/QIODevice.

Also implemented the usage side of it, meaning random key generation
upon uploading new secrets.

I would also like to use this, and the InputStreamQFileAdapter for vmime
parsing, so that I don't need to keep the request in memory, but the
QFileEncrypted doesn't yet do seeking, and that would be required.

Also fixed some bugs in the RequestUploader that I couldn't tie loose
from this commit.